pyproject.toml
[tool.poetry]
name = "louvre"
version = "0.1.0"
description = "Connector to JPG data"
authors = ["SwissQuant"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pandas = "^2.2.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
